feat(app): add GitHub-backed support tickets and providers UI
Adds an in-app Support page backed by Convex + GitHub issues, including background sync and attachments, and improves provider configuration UX (BigModel/GLM as an API-key provider). Also hardens API parsing and fixes multiple Avalonia view/layout issues.

+++ src/KorProxy.Infrastructure/Services/DeviceService.cs
@@ -64,8 +64,8 @@ public sealed class DeviceService : IDeviceService
             ParseDeviceType(d.DeviceType),
             ParsePlatform(d.Platform),
             d.AppVersion,
-            d.LastSeenAt,
-            d.CreatedAt)).ToList();
+            (long)d.LastSeenAt,
+            (long)d.CreatedAt)).ToList();
     }
 
     public async Task<DeviceActionResult> RemoveAsync(string token, string deviceId, CancellationToken ct = default)
@@ -112,6 +112,6 @@ public sealed class DeviceService : IDeviceService
         string DeviceType,
         string Platform,
         string AppVersion,
-        long LastSeenAt,
-        long CreatedAt);
+        double LastSeenAt,
+        double CreatedAt);
 }
